+++ src/TestGenerator/ClassGenerator.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Xepozz\TestIt\TestGenerator;
 
 use Nette\PhpGenerator\ClassType;
+use Nette\PhpGenerator\Method;
 use Nette\PhpGenerator\PhpNamespace;
 use PhpParser\Node\Stmt\Class_;
 use PHPUnit\Framework\TestCase;
@@ -21,11 +22,14 @@ class ClassGenerator
     }
 
     /**
-     * @param \Nette\PhpGenerator\Method[] $methods
-     * @return ClassType
+     * @param Method[] $methods
+     * @return ClassType|null
      */
-    public function generate(array $methods): ClassType
+    public function generate(array $methods): ?ClassType
     {
+        if ($methods === []) {
+            return null;
+        }
         $this->class->setMethods($methods);
        return $this->class;
     }