+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.2.0 under development
 
+- TestGenerator does not generate test classes for classes without methods 
 - Use generators (`yield` statement) instead of returning an array 
 
 ## 1.1.0